+++ src/rogallo/widgets/viewer/gemtext/link.py
@@ -23,6 +23,7 @@ from rich.text import Text
 # Textual imports.
 from textual import on
 from textual.events import Click
+from textual.geometry import Region
 from textual.reactive import reactive, var
 from textual.selection import Selection
 from textual.widget import Widget
@@ -47,6 +48,7 @@ from ....preflight import (
 )
 from ....safe_escape import escape
 from ....types import RogalloLocation, SpartanURINeedingData
+from ..searchable import NEEDLE
 from .content_filter import GemtextContent
 from .icons import icon
 
@@ -58,6 +60,7 @@ class GemtextLink(Widget, can_focus=True):
     COMPONENT_CLASSES = {
         "gemtext-link--icon",
         "gemtext-link--jump-number",
+        NEEDLE,
     }
 
     DEFAULT_CSS = """
@@ -135,6 +138,16 @@ class GemtextLink(Widget, can_focus=True):
         """The normalised URI to use when opening the link."""
         self._filtered_content: Text | str | None = None
         """The filtered content of the link."""
+        self._find_state: int = -1
+        """The current state of the search."""
+        self._needle: str | None = None
+        """The current search needle."""
+
+    def find_reset(self) -> None:
+        """Reset the search state of the widget."""
+        self._find_state = -1
+        self._needle = None
+        self.refresh()
 
     @property
     def normalised_uri(self) -> str:
@@ -230,6 +243,12 @@ class GemtextLink(Widget, can_focus=True):
             link_text.stylize(
                 self.screen.get_component_rich_style("screen--selection", partial=True)
             )
+        if self._needle and self._find_state >= 0:
+            link_text.stylize(
+                self.get_component_rich_style(NEEDLE),
+                self._find_state,
+                self._find_state + len(self._needle),
+            )
         link_data.append(link_text)
 
         # If we're showing link numbers and they're not "cosy".
@@ -258,6 +277,41 @@ class GemtextLink(Widget, can_focus=True):
     def _action_open_link_externally(self) -> None:
         """Open the link in the external browser."""
         open_in_browser(self._normalised_uri)
+
+    def find_next_text(self, needle: str) -> bool:
+        """Find the next occurrence of the needle in the Gemtext list item.
+
+        Args:
+            needle: The string to find.
+
+        Returns:
+            True if the needle was found, False otherwise.
+        """
+        if self._filtered_content is None:
+            return False
+        self._needle = needle
+        self._find_state = (
+            (
+                Text(self._filtered_content)
+                if isinstance(self._filtered_content, str)
+                else self._filtered_content.copy()
+            )
+            .plain.casefold()
+            .find(
+                needle.casefold(),
+                self._find_state + 1 if self._find_state is not None else 0,
+            )
+        )
+        self.refresh()
+        return self._find_state >= 0
+
+    def found_region(self) -> Region:
+        """Get the region of the found text in the widget.
+
+        Returns:
+            The region of the found text, or the widget's region.
+        """
+        return self.virtual_region_with_margin
 
     def get_selection(self, selection: Selection) -> tuple[str, str] | None:
         return selection.extract(f"=> {self.normalised_uri} {self._link}"), "\n"