+++ src/rogallo/widgets/viewer/plain_text.py
@@ -1,0 +1,108 @@
+"""Provides a widget for displaying plain text content."""
+
+##############################################################################
+# Rich imports.
+from rich.text import Text
+
+##############################################################################
+# Textual imports.
+from textual.content import Content
+from textual.geometry import Region
+from textual.style import Style
+from textual.widgets import Static
+
+##############################################################################
+# Local imports.
+from .searchable import NEEDLE
+
+
+##############################################################################
+class PlainText(Static):
+    """A widget for displaying plain text content."""
+
+    COMPONENT_CLASSES = {NEEDLE}
+
+    def __init__(self, text: str | Text | Content, classes: str | None = None) -> None:
+        """Initialise the object.
+
+        Args:
+            text: The plain text to display.
+        """
+        self._text: Content
+        """The content of the plain text."""
+        if isinstance(text, str):
+            self._text = Content(text)
+        elif isinstance(text, Text):
+            self._text = Content.from_rich_text(text)
+        else:
+            self._text = text
+        self._find_state: int = -1
+        """The current state of the search."""
+        super().__init__(text, markup=False, classes=classes)
+
+    def find_reset(self) -> None:
+        """Reset the search state of the widget."""
+        self._find_state = -1
+        self.update(self._text)
+
+    def find_next_text(self, needle: str) -> bool:
+        """Find the next occurrence of the needle in the Gemtext text.
+
+        Args:
+            needle: The string to find.
+
+        Returns:
+            True if the needle was found, False otherwise.
+        """
+        self._find_state = self._text.plain.casefold().find(
+            needle.casefold(),
+            self._find_state + 1 if self._find_state is not None else 0,
+        )
+        if self._find_state < 0:
+            self.update(self._text)
+            return False
+        self.update(
+            self._text.stylize(
+                Style.combine(
+                    (
+                        Style.from_rich_style(self.get_component_rich_style(NEEDLE)),
+                        Style.from_meta({NEEDLE: True}),
+                    )
+                ),
+                self._find_state,
+                self._find_state + len(needle),
+            )
+        )
+        return True
+
+    def found_region(self) -> Region:
+        """Get the region of the found text in the widget.
+
+        Returns:
+            The region of the found text, or the widget's region.
+        """
+        if self._find_state < 0:
+            return self.virtual_region_with_margin
+        matching_lines: list[int] = []
+        gather_match = matching_lines.append
+        for line_number, line in enumerate(
+            self.visual.to_strips(
+                self, self.visual, self.size.width, self.size.height, Style()
+            )
+        ):
+            for segment in line:
+                if segment.style and segment.style.meta.get(NEEDLE) is True:
+                    gather_match(line_number)
+        return (
+            Region(
+                0,
+                self.virtual_region_with_margin.y + matching_lines[0],
+                self.size.width,
+                max(1, len(matching_lines)),
+            )
+            if matching_lines
+            else self.virtual_region_with_margin
+        )
+
+
+### plain_text.py ends here